+++ modules/apps/foundation/vulcan/vulcan-api/src/main/java/com/liferay/vulcan/pagination/SingleModel.java
@@ -14,35 +14,25 @@
 
 package com.liferay.vulcan.pagination;
 
-import com.liferay.vulcan.identifier.Identifier;
-
 /**
  * Provides a wrapper over a model for skipping problems related to the Java
  * generics system.
  *
  * @author Alejandro Hernández
+ * @review
  */
 public class SingleModel<T> {
 
-	public SingleModel(T model, Class<T> modelClass, Identifier identifier) {
+	public SingleModel(T model, Class<T> modelClass) {
 		_model = model;
 		_modelClass = modelClass;
-		_identifier = identifier;
-	}
-
-	/**
-	 * Returns the identifier.
-	 *
-	 * @return the identifier.
-	 */
-	public Identifier getIdentifier() {
-		return _identifier;
 	}
 
 	/**
 	 * Returns the model.
 	 *
 	 * @return the model.
+	 * @review
 	 */
 	public T getModel() {
 		return _model;
@@ -52,12 +42,12 @@ public class SingleModel<T> {
 	 * Returns the model class.
 	 *
 	 * @return the model class.
+	 * @review
 	 */
 	public Class<T> getModelClass() {
 		return _modelClass;
 	}
 
-	private final Identifier _identifier;
 	private final T _model;
 	private final Class<T> _modelClass;